removed duplicated type for poster categories and moved sidebarSections to valSchemas
